reorganize examples test suite
