+++ examples/cube_qt.py
@@ -1,6 +1,7 @@
 """
 Example showing a single geometric cube.
 """
+# run_example = false
 
 import pygfx as gfx
